Simplify codenarc rules for package name.
We're already doubling up the PackageName rule: one which does the
default check, and one which checks if the package name starts with
edu.oregonstate.mist in addition to checking if the package name is
lowercase. Simplify the latter by just checking if the package name
starts with edu.oregonstate.mist.