Fix time zones
The API is supposed to return UTC timestamps, but we were storing local
times in the database, because SYSDATE returns the current local time.

We could change the column datatype from DATE (which does not store time
zone information) to TIMESTAMP WITH TIME ZONE (which does) but we don't
actually care what the local time was. We just want everything to be
UTC.

So use `cast(SYSTIMESTAMP at time zone 'UTC' as DATE)` to get UTC time
as a DATE and store that.