Fix crash when album attributes are missing in POST or PUT request.
CO-454

This takes a little explanation. The crash occurs when trying to
call .intValue() on a null value. The null values aren't the problem.
They represent missing fields in the album object which the user
was trying to PUT or POST. That's fine. The problem is the intValue
calls.

Groovy's type checking seems to have a corner case involving overloaded
methods and implicit casting. SQLStatement.bind() is overloaded to allow
lots of different types of values to be bound. Including Integer and
int. Now, if you try to pass an Integer to a function expecting int,
groovy is smart enough to automatically unwrap it for you. However, if
you try to pass an Integer to bind(), groovy can't decide whether to
call bind(Integer) or bind(int) and complains about an ambigous method
call. There is seemingly no way to tell it that yes, you really meant to
pass an Integer and there's no need to unwrap the value. When I searched
for ways around this, one of the suggestions was to cast the Integer to
an int by calling intValue(). For some reason this worked, so I changed
the code and didn't think about it twice. Of course, this isn't correct
if your Integer can be null.

The fix in this case is to give up trying to get the type right and just
cast the value to Object and let bind() sort it out.